// backend/src/models/ClinicalTrialStudiesResponse.ts
export interface ClinicalTrialStudiesResponse {
  totalCount: number;
  studies: ClinicalTrialStudy[];
  nextPageToken?: string;
}

export interface ClinicalTrialStudy {
  protocolSection: ProtocolSection;
  derivedSection?: DerivedSection;
  hasResults?: boolean;
}

export interface ProtocolSection {
  identificationModule: IdentificationModule;
  statusModule?: StatusModule;
  sponsorCollaboratorsModule?: SponsorCollaboratorsModule;
  oversightModule?: OversightModule;
  descriptionModule?: DescriptionModule;
  conditionsModule?: ConditionsModule;
  designModule?: DesignModule;
  armsInterventionsModule?: ArmsInterventionsModule;
  outcomesModule?: OutcomesModule;
  eligibilityModule?: EligibilityModule;
  contactsLocationsModule?: ContactsLocationsModule;
  referencesModule?: ReferencesModule;
}
// ...
export interface DerivedSection {
  miscInfoModule?: {
    versionHolder?: string;
  };
  conditionBrowseModule?: ConditionBrowseModule;
}